// argefrontend/src/app/views/+basvuru/forms/model/gonullu-popilasyon.ts
/* Sağlıklı gönüllüler
Hastalar
Özel hassas popülasyonlar
Gebe kadınlar
Emziren kadınlar
Acil olgular
Şahsen olur veremeyecek gönüllüler
 */

export class GonulluPopilasyon {
    saglikliGonulluler?: string;
    Hastalar?: string;
    OzelHassasPopilasyonlar?: string;
    GebeKadinlar?: string;
    EmzirenKadinlar?: string;
    AcilOlgular?: string;
    SahsenOlurVeremiyecekler?: string;
    Aciklama?: string;



}